+++ demo/merge-document-basic.php
@@ -5,6 +5,7 @@ include_once __DIR__ . '/bootstrap.php';
 
 use TxTextControl\ReportingCloud\ReportingCloud;
 use TxTextControl\ReportingCloud\Stdlib\ConsoleUtils;
+use TxTextControl\ReportingCloud\Stdlib\FileUtils;
 use TxTextControl\ReportingCloud\Stdlib\Path;
 
 // Instantiate ReportingCloud, using your API key
@@ -37,7 +38,7 @@ $arrayOfBinaryData = $reportingCloud->mergeDocument(
 
 // Write the document's binary data to disk
 
-file_put_contents($destinationFilename, $arrayOfBinaryData[0]);
+FileUtils::write($destinationFilename, $arrayOfBinaryData[0]);
 
 // Output to console the location of the generated document